requirements.txt
uvicorn==0.14.0
fastapi==0.66.0
pydantic==1.8.2
SQLAlchemy
psycopg2
requests
asyncio
typing
pyjwt
bcrypt
passlib
orjson
overpy
aiohttp
python-dotenv